layernorm function test
